you can run the code now

+++ Simulation Framework For CRNs/Performance.cpp
@@ -2,7 +2,7 @@
 
 
 
-Performance::Performance(int timeSlot)
+Performance::Performance(double timeSlot)
 {
 	timeS = timeSlot;
 }
@@ -10,12 +10,12 @@ Performance::Performance(int timeSlot)
 
 void Performance::outputFAFile(const std::vector<int> &PFAvsSU)
 {	
-	int tS = timeS;
-	std::vector<int> oFile;
+	double tS = timeS;
+	std::vector<double> oFile;
 	std::transform(PFAvsSU.begin(), PFAvsSU.end(), std::back_inserter(oFile),
 		[tS](int num) {return num / tS;}); //this function divide all element in vector
 	std::ofstream outputFileFAvsSUid;
 	outputFileFAvsSUid.open("PFA_VS_SUID_FOR_ALLBANDs.csv");
-	std::ostream_iterator<int> outputIterator(outputFileFAvsSUid, "\n");
+	std::ostream_iterator<double> outputIterator(outputFileFAvsSUid, "\n");
 	std::copy(oFile.begin(), oFile.end(), outputIterator);
 }
